requirements.txt
requests
beautifulsoup4
transformers
huggingface-hub
python-dotenv

